containers/images/pytorch-cuda12.1-mpi4.1/docker_wrapper_extensions.py
"""
DockerWrapper extension for PyTorch + CUDA + MPI container.

This module extends the docker-wrapper library with HPC-specific
functionality for PyTorch distributed training workloads.
"""

# When docker-wrapper is available, import from it
# from docker_wrapper import DockerImage

# For now, provide a minimal implementation
class PyTorchHPCImage:
    """PyTorch HPC Docker image extension."""

    def __init__(self, name: str = "pytorch-cuda12.1-mpi4.1"):
        """
        Initialize PyTorch HPC image.

        Args:
            name: Image name/tag
        """
        self.name = name
        self.cuda_version = "12.1"
        self.pytorch_version = "2.1.0"
        self.mpi_version = "4.1.4"

    def get_build_args(self) -> dict:
        """Get build arguments for Docker build."""
        return {
            "PYTORCH_VERSION": self.pytorch_version,
            "CUDA_VERSION": self.cuda_version,
            "MPI_VERSION": self.mpi_version,
        }

    def get_run_args(self, gpu: bool = True) -> dict:
        """
        Get runtime arguments for Docker run.

        Args:
            gpu: Whether to enable GPU support

        Returns:
            Dictionary of runtime arguments
        """
        args = {
            "environment": {
                "PYTHONUNBUFFERED": "1",
                "OMP_NUM_THREADS": "1",
            },
            "volumes": {
                "/workspace": {"bind": "/workspace", "mode": "rw"}
            }
        }

        if gpu:
            args["device_requests"] = [
                {
                    "count": -1,  # All GPUs
                    "capabilities": [["gpu"]],
                }
            ]

        return args

    def test_pytorch_cuda(self) -> str:
        """Get test command for PyTorch CUDA functionality."""
        return """python3 -c "
import torch
print(f'PyTorch: {torch.__version__}')
print(f'CUDA Available: {torch.cuda.is_available()}')
if torch.cuda.is_available():
    print(f'CUDA Version: {torch.version.cuda}')
    print(f'GPU Count: {torch.cuda.device_count()}')
    print(f'GPU Name: {torch.cuda.get_device_name(0)}')
"
"""

    def test_mpi(self) -> str:
        """Get test command for MPI functionality."""
        return """python3 -c "
from mpi4py import MPI
comm = MPI.COMM_WORLD
print(f'MPI Rank: {comm.Get_rank()}')
print(f'MPI Size: {comm.Get_size()}')
"
"""


# Configuration for CMake integration
CONTAINER_CONFIG = {
    "name": "pytorch-cuda12.1-mpi4.1",
    "dockerfile_dir": "Docker",
    "tag": "latest",
    "description": "PyTorch + CUDA 12.1 + MPI 4.1 for HPC",
    "requires_gpu": True,
    "test_commands": [
        "python3 --version",
        "python3 -c 'import torch; print(torch.__version__)'",
        "mpirun --version",
    ]
}
