containers/requirements.txt
# Container build system requirements
# Uses docker-wrapper as external Python library

# Core container management (assuming docker-wrapper is available from PyPI or git)
# docker-wrapper>=1.0.0  # Uncomment when docker-wrapper is published

# For now, using docker-py directly as docker-wrapper alternative
docker>=6.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
jinja2>=3.0.0

# HPC-specific dependencies
# These extend docker-wrapper functionality for HPC workloads
mpi4py>=3.1.0
numpy>=1.24.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
